Remove cheatnet dep from sncast
