+++ crates/forge-runner/src/running.rs
@@ -34,8 +34,10 @@ use cheatnet::runtime_extensions::forge_runtime_extension::{
     get_all_execution_resources, ForgeExtension, ForgeRuntime,
 };
 use cheatnet::runtime_extensions::io_runtime_extension::IORuntimeExtension;
-use cheatnet::state::{BlockInfoReader, CheatnetBlockInfo, CheatnetState, ExtendedStateReader};
+use cheatnet::state::{BlockInfoReader, CheatnetState, ExtendedStateReader};
 use itertools::chain;
+use runtime::context;
+use runtime::context::BlockInfo;
 use runtime::{ExtendedRuntime, StarknetRuntime};
 use starknet::core::utils::get_selector_from_name;
 use starknet_api::core::PatriciaKey;
@@ -131,9 +133,9 @@ pub(crate) fn run_fuzz_test(
     })
 }
 
-fn build_context(block_info: CheatnetBlockInfo) -> EntryPointExecutionContext {
-    let block_context = cheatnet_constants::build_block_context(block_info);
-    let account_context = cheatnet_constants::build_transaction_context();
+fn build_context(block_info: BlockInfo) -> EntryPointExecutionContext {
+    let block_context = context::build_block_context(block_info);
+    let account_context = context::build_transaction_context();
 
     EntryPointExecutionContext::new(
         &block_context,
