+++ crates/sncast/Cargo.toml
@@ -22,7 +22,6 @@ toml = "0.8.8"
 rpassword = "7.3.1"
 promptly = "0.3.1"
 scarb-api = { path = "../scarb-api" }
-cheatnet = { path = "../cheatnet" }
 # Freeze reqwest dependency, because newer versions cause cast tests to fail on macs
 reqwest = "=0.11.18"
 indoc.workspace = true
